requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
flask==3.0.0
flask-cors==4.0.0
selenium==4.17.2
opencv-python==4.9.0.80
pillow==10.2.0
numpy==1.26.3
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-multipart==0.0.6
pynput==1.7.6
pytesseract==0.3.10
undetected-chromedriver==3.5.4
webdriver-manager==4.0.1
pytest==7.4.4
pytest-asyncio==0.23.3
aiofiles==23.2.1
python-dotenv==1.0.0
